Move the box solver to src/internal/solver.ts
LayoutNode and the flex, table, grid, block and margin-collapse
algorithms over it know nothing about the DOM, and flex.test already
drives them without one. layout.ts keeps the mapping from the DOM to
boxes and imports the solver.
